Move /internal/bpf2go to /bpf2go to make it externally importable (I love Go)
