+++ bpf2go/output_test.go
